requirements-upgrade.txt
# AET System Upgrade Dependencies
# Additional packages needed for the upgrade phases

# Phase 0: Critical Fixes
structlog==24.1.0           # Structured logging

# Phase 1: Quick Wins  
psutil==5.9.8               # Resource monitoring

# Phase 2: State Recovery
# (Uses existing packages)

# Phase 3: Simple Mode
# (Uses existing packages)

# Phase 4: Observability
prometheus-client==0.19.0    # Prometheus metrics
opentelemetry-api==1.22.0   # Tracing API
opentelemetry-sdk==1.22.0   # Tracing SDK
opentelemetry-exporter-otlp==1.22.0  # OTLP exporter
opentelemetry-instrumentation-flask==0.43b0  # Flask instrumentation
opentelemetry-instrumentation-requests==0.43b0  # Requests instrumentation

# Phase 4: Production Deployment
gunicorn==21.2.0            # Production WSGI server

# Testing (All Phases)
pytest==7.4.4               # Test framework
pytest-cov==4.1.0           # Coverage reporting
pytest-asyncio==0.23.3      # Async test support
pytest-benchmark==4.0.0      # Performance benchmarking
pytest-mock==3.12.0         # Mocking support

# Development Tools
black==23.12.1              # Code formatting
flake8==7.0.0              # Linting
mypy==1.8.0                # Type checking

# Existing dependencies (for reference)
# These should already be installed:
# flask
# requests
# sqlite3 (built-in)
# sentence-transformers (optional for KM semantic search)